ingest/requirements.txt
psycopg2-binary
requests
python-dotenv
pandas
pyyaml
openmeteo-requests
requests-cache
retry-requests
numpy
sodapy
pytest